+++ lib/sanbase/external_services/coinmarketcap/ticker_fetcher.ex
@@ -63,7 +63,14 @@ defmodule Sanbase.ExternalServices.Coinmarketcap.TickerFetcher do
         market_cap_usd: ticker.market_cap_usd,
 	name: ticker.name,
 	price_usd: ticker.price_usd,
+  rank: ticker.rank,
+  volume_usd: ticker.'24h_volume_usd',
+  available_supply: ticker.available_supply,
+  total_supply: ticker.total_supply,
 	symbol: ticker.symbol,
+  percent_change_1h: ticker.percent_change_1h,
+  percent_change_24h: ticker.percent_change_24h,
+  percent_change_7d: ticker.percent_change_7d,
 	update_time: DateTime.from_unix!(ticker.last_updated)
       })
     |> Repo.insert_or_update!